+++ splitcal/splitcalHit.h
@@ -43,16 +43,11 @@ class splitcalHit : public ShipHit
     void SetIsY(bool y) {_isY = y;}
     void SetIsUsed(int u) {_isUsed = u;}
     void SetXYZErrors(double xError, double yError, double zError) {_xError = xError; _yError = yError; _zError = zError;}
-    /* void SetClusterIndex(int i) {_clusterIndex = i;} */
-    /* void SetEnergyWeight(double w) {_energyWeight = w;} */
-    void AddClusterIndex(int i) {_vecClusterIndices.push_back(i);}
-    void AddEnergyWeight(double w) {_vecEnergyWeights.push_back(w);}
 
     double GetX() const { return _x; }
     double GetY() const { return _y; }
     double GetZ() const { return _z; }
     double GetEnergy() const { return _energy; }
-    double GetEnergyForCluster(int i) const;
     int GetIsPrecisionLayer() const { return _isPrecisionLayer; }
     int GetLayerNumber() const { return _nLayer; }
     int GetModuleXNumber() const { return _nModuleX; }
@@ -64,12 +59,6 @@ class splitcalHit : public ShipHit
     double GetXError() const { return _xError; }
     double GetYError() const { return _yError; }
     double GetZError() const { return _zError; }
-    /* double GetEnergyWeight() {return _energyWeight;} */
-    /* int GetClusterIndex() {return _clusterIndex;} */
-    std::vector<int> GetClusterIndices() const { return _vecClusterIndices; }
-    std::vector<double> GetEnergyWeights() const { return _vecEnergyWeights; }
-    bool IsShared() const { return GetClusterIndices().size() > 1; }
-    double GetEnergyWeightForIndex(int index) const;
 
     /** Copy constructor **/
     splitcalHit(const splitcalHit& point) = default;
@@ -79,14 +68,11 @@ class splitcalHit : public ShipHit
 
     Float_t flag;
     double _x, _y, _z, _xError, _yError, _zError;
-    double _energy;// _energyWeight;
-    /* std::string _nameSubDetector;  */
-    int _isPrecisionLayer, _nLayer, _nModuleX, _nModuleY, _nStrip, _isUsed;// _clusterIndex;
+    double _energy;
+    int _isPrecisionLayer, _nLayer, _nModuleX, _nModuleY, _nStrip, _isUsed;
     bool _isX, _isY;
-    std::vector<double > _vecEnergyWeights;
-    std::vector<int > _vecClusterIndices;
 
-    ClassDef(splitcalHit,3);
+    ClassDef(splitcalHit,4);
 
 
 };